feat: 공유 일정 검색 UI 구조 정리
