+++ src/shared/constants/scheduleShareFriend.ts
@@ -1,0 +1,12 @@
+import type { ScheduleShareFriend } from '@/shared/types/schedule/shareFriend'
+
+// TODO: 친구 목록 API 연동 필요
+export const MOCK_SCHEDULE_SHARE_FRIENDS: ScheduleShareFriend[] = [
+  { userName: '김철수', userId: '1', email: 'asdf@asdf.com' },
+  { userName: '김철수', userId: '5', email: 'asdf@asdf.com' },
+  { userName: '김철수', userId: '6', email: 'asdf@asdf.com' },
+  { userName: '김철수', userId: '7', email: 'asdf@asdf.com' },
+  { userName: '이영희', userId: '2', email: 'asdf@asdf.com' },
+  { userName: '박민수', userId: '3', email: 'asdf@asdf.com' },
+  { userName: '최지우', userId: '4', email: 'asdf@asdf.com' },
+]
