+++ src/shared/types/modal/scheduleEditor.ts
@@ -14,6 +14,7 @@ export type ScheduleEditorFormProps = {
   eventId: CalendarEvent['id']
   onClose: () => void
   isEditing?: boolean
+  isShared?: boolean
   headerTitlePortalTarget?: HTMLElement | null
   initialEvent?: CalendarEvent | null
   onEventColorChange?: (eventId: CalendarEvent['id'], color: EventColorType) => void
