+++ src/shared/types/schedule/shareFriend.ts
@@ -1,0 +1,5 @@
+export type ScheduleShareFriend = {
+  userName: string
+  userId: string
+  email: string
+}
